Guess there were a few things left to do before 1.0 ;)
Renamed and rearranged a couple of classes/files. It's much nicer now.

Added extra/basicmenu.xpp as a base for the scenes and handlers used for
SFGUI menus.

Fixed issue with using multiple SFGUI Menus. A small side effect of this
is that SFGUI is now a dependency of libsqee/base, rather than
libsqee/extra, but this really isn't an issue.

Added stubs for an ingame (pause) menu to the sokoban test.