Last major commit for probably a few months, as I'm going to start on a World Editor.
Added the sponza atrium. It looks neat.
Added support for a binary sqee mesh format and converted some meshes to it.
Added support of jpg textures, and converted some big textures to jpg to save space.
Added ModelStatic and ModelSkelly classes to libsqee.
Removed a lot of the old Heightlayers code.
Rewrote the Skeleton and Animation classes. They are faster now, with nicer interfaces.
Lots of other stuff. This release is still pretty messy, but oh well. I need to start on the editor.